backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
bcrypt==5.0.0
PyMySQL==1.1.1
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.28.1
